% Auto-generated: do not edit by hand
\name{DashNglDashNgl}

\alias{DashNglDashNgl}

\title{DashNgl component}

\description{
Dash ngl component
}

\usage{
DashNglDashNgl(id=NULL, data=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{data}{List of lists containing elements 'filename', 'ext', 'config'.
those elements have the following types:
  - filename (character; required)
  - ext (character; optional)
  - config (optional): . config has the following type: lists containing elements 'type', 'input'.
those elements have the following types:
  - type (character; required)
  - input (unnamed list | named list | character; optional)s | lists containing elements 'filename', 'ext', 'config'.
those elements have the following types:
  - filename (character; required)
  - ext (character; optional)
  - config (optional): . config has the following type: lists containing elements 'type', 'input'.
those elements have the following types:
  - type (character; required)
  - input (unnamed list | named list | character; optional). Custom property}
}
